Refactor executable loading interface
